Refactor player-related layout and routing structure
Introduced `PlayerLayout` to manage shared layout and context for
player-related pages, simplifying components like `StartedHunt`
and `HuntPage`. Refactored routing in `main.jsx` for better
organization. Removed redundant logic and unused imports,
improving modularity and maintainability.
