+++ scavengerhunt.react/src/pages/HuntPage.jsx
@@ -1,46 +1,10 @@
-import React from 'react';
 import { Center, Title } from '@mantine/core';
-import { useNavigate, useParams } from 'react-router';
+import { useOutletContext } from 'react-router';
 
-import { useAlertDispatch } from '../utils/AlertContext';
-import { getHunt } from '../services/huntService';
-import { isPlayerValid } from '../services/playerService';
 import StartedHunt from '../components/StartedHunt';
 
 export default function HuntPage() {
-  const [hunt, setHunt] = React.useState();
-
-  const alertDispatch = useAlertDispatch();
-  const navigate = useNavigate();
-  const params = useParams();
-
-  React.useEffect(() => {
-    async function getHuntDetails() {
-      try {
-        const hunt = await getHunt(params.huntid, undefined);
-        setHunt(hunt);
-      }
-      catch (err) {
-        console.error('Failed to get hunt details', err);
-        alertDispatch({ type: 'error', message: err, show: true });
-      }
-    }
-
-    async function getPlayerDetails() {
-      try {
-        const isValid = await isPlayerValid();
-        if (!isValid) throw new Error('You are not a valid player. Please join again');
-      }
-      catch (err) {
-        console.error('Not a valid player', err);
-        navigate('/hunt/join');
-        alertDispatch({ type: 'error', message: err.message, show: true });
-      }
-    }
-
-    getHuntDetails();
-    getPlayerDetails();
-  }, []);
+  const { hunt } = useOutletContext();
 
   return (
     <>
